add:SPI实现类添加注释，实现protostuff序列化方式

+++ simple-Qrpc-framework/src/main/java/com/siman/qrpc/registry/zk/util/CuratorUtils.java
@@ -13,6 +13,7 @@ import org.apache.curator.framework.recipes.cache.PathChildrenCacheListener;
 import org.apache.curator.retry.ExponentialBackoffRetry;
 import org.apache.zookeeper.CreateMode;
 
+import java.net.InetSocketAddress;
 import java.util.List;
 import java.util.Map;
 import java.util.Properties;
@@ -112,10 +113,12 @@ public final class CuratorUtils {
      * 清空注册中心
      * @param zkClient
      */
-    public static void clearRegistry(CuratorFramework zkClient) {
-        REGISTERED_PATH_SET.stream().parallel().forEach(p -> {
+    public static void clearRegistry(CuratorFramework zkClient, InetSocketAddress inetSocketAddress) {
+        REGISTERED_PATH_SET.stream().forEach(p -> {
             try {
-                zkClient.delete().forPath(p);
+                if (p.endsWith(inetSocketAddress.toString())) {
+                    zkClient.delete().forPath(p);
+                }
             } catch (Exception exception) {
                 log.error("清除节点:[{}] 失败", p);
             }
